requirements.txt
groq>=0.13.0
google-generativeai>=0.8.0
resend>=2.23.0
requests>=2.31.0
python-dotenv>=1.0.0
jinja2>=3.1.0
